Move device factory mock to a test module
